Split off a separate utility for CLI versions
